refactor: extract TeamRecordSectionRenderer; wire into RecordHoldersView
